+++ CHANGES.md
@@ -6,6 +6,7 @@ All notable changes grouped by version.
 
  - `BC-BREAK` Now DNS lookup limit is enforced as 10 maximum queries for mechanisms and 10 maximum queries for retrieving A record out of exchanges returned in the MX mechanism.
  - New option `{ maxDNS: 10 }` to change the DNS lookup limit.
+ - `BC-BREAK` Domain and sender parameters has been moved from `SPF.check()` to `SPF.constructor()`. Stable API has no change.
 
 ## [0.2.0] 2018-12-25